% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-preprocessing.R
\name{add_angle_difference}
\alias{add_angle_difference}
\title{Adds new colum angle_diff_axis where it calculates angle difference between rows}
\usage{
add_angle_difference(player_log, axis = "x")
}
\arguments{
\item{player_log}{data.table log that gets modified by reference}

\item{axis}{axis X, Y or Z. the player log shoudl have Rotation.Axis colum}
}
\description{
Adds new colum angle_diff_axis where it calculates angle difference between rows
}
