% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-modifying.R
\name{add_goal_order.brainvr}
\alias{add_goal_order.brainvr}
\title{Adds goal order vector to determine in what order goals came}
\usage{
add_goal_order.brainvr(obj, order)
}
\arguments{
\item{obj}{Brainvr object}

\item{order}{vector of goal order. eg. (1, 3, 5, 1)}
}
\value{
Brainvr object with added field
}
\description{
Adds goal order vector to determine in what order goals came
}
