% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-modifying.R
\name{add_goal_positions.brainvr}
\alias{add_goal_positions.brainvr}
\title{Loads goal positions to the Brainvr object}
\usage{
add_goal_positions.brainvr(obj, df)
}
\arguments{
\item{obj}{Brainvr object}

\item{df}{data.frame with goal positions. Goal row must correspond to goal order}
}
\description{
Loads goal positions to the Brainvr object
}
