% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-visualising.R
\name{add_limits}
\alias{add_limits}
\title{Checks if object has map limits variable and adds plot limits if so}
\usage{
add_limits(plt, obj)
}
\arguments{
\item{plt}{existing plot}

\item{obj}{Brainvr object}
}
\description{
Checks if object has map limits variable and adds plot limits if so
}
