% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-calculations.R
\name{angle_from_positions}
\alias{angle_from_positions}
\title{Zero vector is the vector given by the unity for calculation rotation of oabjects. 
In 3D unity (different in UNREAL!) are exes X horizontal left/right, Y vertical up/down, and Z plane horizontal up/down
The zero vector is vector when GameObject only changes its Z position, therefore if we calculate position as [X Z], 
the normalised zero vector shoudl be 0 on X and 1 on Z - [0, 1]}
\usage{
angle_from_positions(pos_from, pos_to)
}
\description{
Zero vector is the vector given by the unity for calculation rotation of oabjects. 
In 3D unity (different in UNREAL!) are exes X horizontal left/right, Y vertical up/down, and Z plane horizontal up/down
The zero vector is vector when GameObject only changes its Z position, therefore if we calculate position as [X Z], 
the normalised zero vector shoudl be 0 on X and 1 on Z - [0, 1]
}
