% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-visualising.R
\name{brainvr.plot_trial_path}
\alias{brainvr.plot_trial_path}
\title{Plots trial path}
\usage{
brainvr.plot_trial_path(obj, trialId)
}
\arguments{
\item{trialId}{}
}
\value{
ggplot objectg
}
\description{
Plots trial path
}
