% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-visualising.R
\name{brainvr.plot_trials_paths}
\alias{brainvr.plot_trials_paths}
\title{PLots multiple paths}
\usage{
brainvr.plot_trials_paths(obj, columns = 5, indices = c())
}
\arguments{
\item{obj}{}

\item{indices}{}
}
\description{
PLots multiple paths
}
