% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-visualising.R
\name{create_plot}
\alias{create_plot}
\title{Creates empty stylised plot with given constriaints, if set in the object}
\usage{
create_plot(obj)
}
\arguments{
\item{obj}{UnityObject}
}
\value{
ggplot plot with x and y lims if set#'
}
\description{
Creates empty stylised plot with given constriaints, if set in the object
}
