% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-exporting.R
\name{export_position_log}
\alias{export_position_log}
\title{Exports preprocessed player log}
\usage{
export_position_log(log, id, path = getwd())
}
\arguments{
\item{playerLog:}{data table log as preprocessed by the Analysis fuynction}

\item{path:}{path to the file. Defaults to workign directory}
}
\description{
Exports preprocessed player log
}
