% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-getters.R
\name{get_distance_timewindow}
\alias{get_distance_timewindow}
\title{Returns amount of moved distance in given timewindow}
\usage{
get_distance_timewindow(obj, start, end)
}
\arguments{
\item{obj}{brainvr object}

\item{start}{movement start timestamp}

\item{end}{movement end timestamp}
}
\value{
moved distance or NA if the
}
\description{
Returns amount of moved distance in given timewindow
}
