% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-getters.R
\name{get_experiment_log}
\alias{get_experiment_log}
\title{Returns experiment log data}
\usage{
get_experiment_log(obj, ...)
}
\arguments{
\item{obj}{Brainvr object}
}
\value{
Dataframe with all experiment events
}
\description{
Returns experiment log data
}
