% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-getters.R
\name{get_experiment_settings}
\alias{get_experiment_settings}
\title{Returns experiment settings in a list}
\usage{
get_experiment_settings(obj)
}
\arguments{
\item{obj}{Brainvr Object}
}
\value{
List with saved experiment settings
}
\description{
Returns experiment settings in a list
}
