% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-getters.R
\name{get_finished_trials}
\alias{get_finished_trials}
\title{Returns which trials were finished}
\usage{
get_finished_trials(obj)
}
\arguments{
\item{obj}{BrainvrObject}
}
\value{
indices of finished trials
}
\description{
Returns which trials were finished
}
