% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-getters.R
\name{get_goal_position.brainvr}
\alias{get_goal_position.brainvr}
\title{Returns position of goal at particular index}
\usage{
get_goal_position.brainvr(obj, index, onlyXY = T)
}
\arguments{
\item{obj}{brainvr object}

\item{index}{Goal index}

\item{onlyXY}{if only horizontal coordinates shoud be returned}
}
\description{
Returns position of goal at particular index
}
