% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-getters.R
\name{get_log}
\alias{get_log}
\title{Returns entire player log}
\usage{
get_log(obj, ...)
}
\arguments{
\item{obj}{Brainvr object}
}
\value{
player log data.frame
}
\description{
Returns entire player log
}
