% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-getters.R
\name{get_log_timewindow.brainvr}
\alias{get_log_timewindow.brainvr}
\title{Returns log between designated times}
\usage{
get_log_timewindow.brainvr(obj, start, end)
}
\arguments{
\item{obj}{}

\item{start}{}

\item{end}{}
}
\description{
Returns log between designated times
}
