% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-getters.R
\name{get_player_log_trial}
\alias{get_player_log_trial}
\title{Returns data.table with player log for particular trial}
\usage{
get_player_log_trial(obj, trialId)
}
\value{
player log for particulat trial
}
\description{
Returns data.table with player log for particular trial
}
