% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-getters.R
\name{get_position_timewindow}
\alias{get_position_timewindow}
\title{Returns log between designated times}
\usage{
get_position_timewindow(obj, start, end, ...)
}
\arguments{
\item{end}{}
}
\value{
navr object with preprocessed times
}
\description{
Returns log between designated times
}
