% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-getters.R
\name{get_results_log}
\alias{get_results_log}
\title{Return results log data}
\usage{
get_results_log(obj, ...)
}
\arguments{
\item{obj}{Brainvr object with loaded results}

\item{...}{}
}
\value{
results data.frame or NULL if empty
}
\description{
Return results log data
}
