% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-getters.R
\name{get_trial_distance.brainvr}
\alias{get_trial_distance.brainvr}
\title{Returns walked distance in a particular trial}
\usage{
\method{get_trial_distance}{brainvr}(obj, trialId)
}
\arguments{
\item{obj}{}

\item{trialId}{}
}
\description{
Returns walked distance in a particular trial
}
