% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-getters.R
\name{get_trial_distance}
\alias{get_trial_distance}
\title{Returns walked distance in a particular trial}
\usage{
get_trial_distance(obj, iTrial)
}
\arguments{
\item{obj}{Brainvr object}

\item{iTrial}{trial index}
}
\description{
Returns walked distance in a particular trial
}
