% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-getters.R
\name{get_trial_duration.brainvr}
\alias{get_trial_duration.brainvr}
\title{Returns how long the trial took and removes potential pauses in the log}
\usage{
\method{get_trial_duration}{brainvr}(obj, trialId, without_pauses = T,
  pause_limit = 10, path_limit = 2)
}
\arguments{
\item{obj}{BrainVr object with preprocessed player log}

\item{trialId}{}

\item{without_pauses}{Defaults to true}

\item{pause_limit}{minimum time to be considered pause}

\item{path_limit}{maximum distance to be considered not moving}
}
\description{
Returns how long the trial took and removes potential pauses in the log
}
