% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-generics.R
\name{get_trial_duration}
\alias{get_trial_duration}
\title{Returns how long the trial took and removes potential pauses in the log}
\usage{
get_trial_duration(obj, trialId, without_pauses = T, ...)
}
\arguments{
\item{obj}{}

\item{without_pauses}{Defaults to true}

\item{...}{}
}
\description{
Returns how long the trial took and removes potential pauses in the log
}
