% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-getters.R
\name{get_trial_event_times}
\alias{get_trial_event_times}
\title{Returns times of certain events happening in particular trials}
\usage{
get_trial_event_times(obj, iTrial, event_name = "")
}
\arguments{
\item{obj}{}

\item{iTrial}{vector or trial indices}
}
\value{
times of certain events
}
\description{
Returns times of certain events happening in particular trials
}
