% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-getters.R
\name{get_trial_log.brainvr}
\alias{get_trial_log.brainvr}
\title{Returns data.table with player log for a particular trial}
\usage{
\method{get_trial_log}{brainvr}(obj, trialId)
}
\value{
player log for particulat trial
}
\description{
Returns data.table with player log for a particular trial
}
