% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-getters.R
\name{get_trial_log}
\alias{get_trial_log}
\title{Returns data.table with player log for a particular trial}
\usage{
get_trial_log(obj, iTrial, ...)
}
\arguments{
\item{obj}{BrainvrObject}

\item{iTrial}{index of the trial to be analyzed}
}
\value{
player log for particulat trial
}
\description{
Returns data.table with player log for a particular trial
}
