% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-getters.R
\name{get_trial_position}
\alias{get_trial_position}
\title{Returns navr_object with data for a particular trial}
\usage{
get_trial_position(obj, iTrial, ...)
}
\arguments{
\item{obj}{BrainvrObject}
}
\value{
navr object
}
\description{
Returns navr_object with data for a particular trial
}
