% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-getters.R
\name{get_trial_times.brainvr}
\alias{get_trial_times.brainvr}
\title{Gets start and finish times of trial}
\usage{
\method{get_trial_times}{brainvr}(obj, trialId)
}
\value{
list with waitingToStart, start and finish
}
\description{
Gets start and finish times of trial
}
