% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-getters.R
\name{get_trial_timewindow}
\alias{get_trial_timewindow}
\title{Gets start and finish times of trial}
\usage{
get_trial_timewindow(obj, trialId)
}
\value{
list with waitingToStart, start and finish
}
\description{
Gets start and finish times of trial
}
