% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-getters.R
\name{get_walked_distnace_timewindow}
\alias{get_walked_distnace_timewindow}
\title{Returns walked distance in given timewindow}
\usage{
get_walked_distnace_timewindow(obj, start, end)
}
\arguments{
\item{obj}{brainvr object}

\item{timeweindow}{}
}
\description{
Returns walked distance in given timewindow
}
