% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-loading.R
\name{load_brainvr_log}
\alias{load_brainvr_log}
\title{Loads a generic brainvr framework log. These logs have specific header and
data notaions.}
\usage{
load_brainvr_log(filepath, func = NULL)
}
\arguments{
\item{filepath}{path to the log}

\item{func}{optional loading function}
}
\value{
list with parsed data and optionally $data field with log's dataframe
}
\description{
Loads a generic brainvr framework log. These logs have specific header and
data notaions.
}
