% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-loading.R
\name{load_experiment_info}
\alias{load_experiment_info}
\title{Loads particular info file into a list}
\usage{
load_experiment_info(filepath)
}
\arguments{
\item{filepath}{path to the file}
}
\value{
list object
}
\description{
Loads particular info file into a list
}
