% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-loading.R
\name{load_experiment_log}
\alias{load_experiment_log}
\title{Loads expeirment log into a predefined list}
\usage{
load_experiment_log(filepath)
}
\arguments{
\item{filepath}{path tot he expeirment log}
}
\value{
list with loaded settings files and data
}
\description{
Loads expeirment log into a predefined list
}
