% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-loading.R
\name{load_header_section}
\alias{load_header_section}
\title{Parses brainvr framework headers}
\usage{
load_header_section(txt)
}
\arguments{
\item{txt}{}
}
\value{

}
\description{
Parses brainvr framework headers
}
