% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-loading.R
\name{load_headers}
\alias{load_headers}
\title{Loads the header portion from the log. Overall simply a wrapper around a 
\code{\link{load_header_section}} function which parses the header from a 
text}
\usage{
load_headers(filepath)
}
\arguments{
\item{filepath}{path to the file}
}
\value{
list of parsed settings
}
\description{
Loads the header portion from the log. Overall simply a wrapper around a 
\code{\link{load_header_section}} function which parses the header from a 
text
}
