% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-loading.R
\name{load_result_log}
\alias{load_result_log}
\title{Loads results log at a specific path}
\usage{
load_result_log(filepath)
}
\arguments{
\item{filepath}{}
}
\value{

}
\description{
Loads results log at a specific path
}
