% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-modifying.R
\name{mirror_axes}
\alias{mirror_axes}
\title{Mirrors X and Y to negative and recomputes rotation}
\usage{
mirror_axes(obj)
}
\value{
UnityObject with mirrored axes and rotation
}
\description{
Mirrors X and Y to negative and recomputes rotation
}
