% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-loading.R
\name{open_experiment_info}
\alias{open_experiment_info}
\title{Searches the directory for experiment log files. Returs single one if multiple are found}
\usage{
open_experiment_info(directory, log_timestamp = NULL, returnSingle = FALSE)
}
\arguments{
\item{directory}{path to the directory where to search}
}
\value{
list with a single loaded info log
}
\description{
Searches the directory for experiment log files. Returs single one if multiple are found
}
