% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-loading.R
\name{open_result_log}
\alias{open_result_log}
\title{Searches for results logs and returns loaded list.}
\usage{
open_result_log(dir, exp_timestamp = NULL)
}
\arguments{
\item{exp_timestamp}{timestamp of a particular results log}

\item{directory}{where to search}
}
\value{
list with loaded result log
}
\description{
REsults log have a _results_ in their filename. 
In case your's doesn't, use load_result_log instead
}
