% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-visualising.R
\name{plot_add_direction}
\alias{plot_add_direction}
\title{Adds arrow pointing from a point in a specified angle}
\usage{
plot_add_direction(plt, position, angle, len)
}
\arguments{
\item{plt}{PLot to which to add the arrow}

\item{position_df}{data.frame. Needs to have columns x, y, angle, length, type}
}
\value{
built ggplot2
}
\description{
Adds arrow pointing from a point in a specified angle
}
