% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-visualising.R
\name{plot_add_player_path}
\alias{plot_add_player_path}
\title{Adds player path to the given plot}
\usage{
plot_add_player_path(plt, df_pos)
}
\arguments{
\item{plot}{}

\item{data}{Data frame with Position.x and Position.z columns}
}
\value{
plot with plotted pat
}
\description{
Adds player path to the given plot
}
