% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-visualising.R
\name{plot_add_points}
\alias{plot_add_points}
\title{Adds specified points to the given plot}
\usage{
plot_add_points(plt, ls)
}
\arguments{
\item{ls}{list with XY vectors. eg. (list(start = c(0, 0), end = C(10, 5)))}

\item{plot}{already created ggplot}
}
\value{
modified plot
}
\description{
Adds specified points to the given plot
}
