% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-visualising.R
\name{plot_trial_path.brainvr}
\alias{plot_trial_path.brainvr}
\title{Plots trial path}
\usage{
\method{plot_trial_path}{brainvr}(obj, trialId)
}
\arguments{
\item{obj}{}

\item{trialId}{}
}
\description{
Plots trial path
}
