% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-visualising.R
\name{plot_trial_path}
\alias{plot_trial_path}
\title{Plots trial path}
\usage{
plot_trial_path(obj, iTrial, ...)
}
\arguments{
\item{obj}{BrainvrObject}

\item{iTrial}{trial index (starting with 1)}
}
\description{
Plots trial path
}
\examples{

}
