% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-preprocessing.R
\name{prepare_navr_log}
\alias{prepare_navr_log}
\title{SUPER IMPORTANT - renames Unity Z to Y and vice versa, because NAVR calculates
speeds from x and y not x and z}
\usage{
prepare_navr_log(position_log)
}
\description{
SUPER IMPORTANT - renames Unity Z to Y and vice versa, because NAVR calculates
speeds from x and y not x and z
}
