% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-preprocessing.R
\name{preprocess_player_log}
\alias{preprocess_player_log}
\title{preprocesses player log and returns if it changed. Passes by referennce}
\usage{
preprocess_player_log(player_log, type = "rigidbody")
}
\arguments{
\item{player_log}{Preprocesses player log by reference and returs if it was changed for saving purposes}

\item{type}{what type of log we have - depending on the brainvr-unity-framework type of player controller used? Possible types are: "rigidbody", "virtualizer"}

\item{player_log}{loaded DATA.TABLE player log}
}
\value{
boolean if changed. Default false
}
\description{
preprocesses player log and returns if it changed. Passes by referennce
}
