% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-modifying.R
\name{resize_layout}
\alias{resize_layout}
\title{Resizes the map to fit the new constraints.}
\usage{
resize_layout(obj, multiplier)
}
\arguments{
\item{obj}{UnityObject}

\item{multiplier}{Conversion of a single unit to a new metric}
}
\value{
ModifiedUnity object
}
\description{
Resizes the map to fit the new constraints.
}
