% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-preprocessing.R
\name{save_preprocessed_player}
\alias{save_preprocessed_player}
\title{Saves preprocessed player ot hte given folder. Receives either specific name, or gets the name from already present player logs
If there are multiple player logs in the folder,takes the name from the first available}
\usage{
save_preprocessed_player(directory, player_log, exp_timestamp = NULL,
  orig_filename = NULL)
}
\arguments{
\item{directory}{Where shoudl i save the file. Should end with a slash!}

\item{player_log}{table to be saved}

\item{orig_filename}{what is the name of the original file}
}
\value{

}
\description{
Saves preprocessed player ot hte given folder. Receives either specific name, or gets the name from already present player logs
If there are multiple player logs in the folder,takes the name from the first available
}
