% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-modifying.R
\name{translate_positions}
\alias{translate_positions}
\title{Moves all positions to correspond to the new 0, 0 coordinate}
\usage{
translate_positions(obj, new_zero)
}
\arguments{
\item{new_zero}{New X and Y vector. Needs to be a int vector of lenght 2 (X, Y)}
}
\description{
Operates on given UnityObject, as it needs to fix not only player log, but potential goal positions in the object as well
}
