% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-preprocessing.R
\name{unity_vector_to_numeric}
\alias{unity_vector_to_numeric}
\title{Converts output of unity loggin a vector 3 to a vector of three}
\usage{
unity_vector_to_numeric(text)
}
\arguments{
\item{text}{unity (x, y, z) vector3 ouput}
}
\description{
Unity logs vector as (x, y, z). This function simply 
splits it into three parts and returns a numeric(2 or 3)
}
