% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-helpers.R
\name{angle_from_positions}
\alias{angle_from_positions}
\title{Creates angle between two points given ZERO vector}
\usage{
angle_from_positions(pos_from, pos_to)
}
\arguments{
\item{pos_from}{position from}
}
\value{
angle in 360
}
\description{
Creates angle between two points given ZERO vector
}
