% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-getters.R
\name{get_trial_times}
\alias{get_trial_times}
\title{Gets start and finish times of trial}
\usage{
get_trial_times(obj, iTrial, ...)
}
\arguments{
\item{obj}{BrainvrObject}

\item{iTrial}{trial index(starting with 1)}
}
\value{
list with waitingToStart, start and end
}
\description{
Gets start and finish times of trial
}
