% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-loading.R
\name{load_experiment}
\alias{load_experiment}
\title{Loads files form a folder into BrainvrObject}
\usage{
load_experiment(folder, exp_timestamp = NULL, override = FALSE)
}
\arguments{
\item{folder}{path to the folder respective to the working directory}
}
\description{
Loads files form a folder into BrainvrObject
}
