% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-loading.R
\name{load_experiments}
\alias{load_experiments}
\title{Goes through the folder and loads every experiment info into separate object}
\usage{
load_experiments(folder, objectFun)
}
\arguments{
\item{folder}{where to look}

\item{objectFun}{what function to create the object that will hold the info}
}
\value{
list of objecs
}
\description{
Goes through the folder and loads every experiment info into separate object
}
