% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-loading.R
\name{open_brainvr_logs}
\alias{open_brainvr_logs}
\title{Generic loading of all the results, experiment and other logs}
\usage{
open_brainvr_logs(directory, log_name, func, exp_timestamp = NULL,
  flatten = F)
}
\arguments{
\item{directory}{}

\item{log_name}{}

\item{exp_timestamp}{}

\item{flatten}{in case only a single file is found, should it be unnested? defaults to false}
}
\value{

}
\description{
Generic loading of all the results, experiment and other logs
}
