% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-loading.R
\name{open_experiment_infos}
\alias{open_experiment_infos}
\title{Searches the directory for experiment log files. Returs single one if multiple are found}
\usage{
open_experiment_infos(directory, log_timestamp = NULL)
}
\arguments{
\item{directory}{path to the directory where to search}

\item{log_timestamp}{}
}
\value{
list with a single loaded info log
}
\description{
Searches the directory for experiment log files. Returs single one if multiple are found
}
