% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-visualising.R
\name{plot_trial_path.brainvr}
\alias{plot_trial_path.brainvr}
\title{Plots brainvr object path}
\usage{
plot_trial_path.brainvr(obj, trialId)
}
\arguments{
\item{obj}{}

\item{trialId}{}
}
\description{
Plots brainvr object path
}
