% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-preprocessing.R
\name{preprocess_player_log}
\alias{preprocess_player_log}
\title{Preprocesses player log and returns it}
\usage{
preprocess_player_log(player_log)
}
\arguments{
\item{player_log}{loaded DATA.TABLE player log}
}
\value{
preprocessed player log
}
\description{
Preprocesses player log and returns it
}
