% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-modifying.R
\name{translate_positions}
\alias{translate_positions}
\title{Moves all positions to correspond to the new 0, 0 coordinate}
\usage{
translate_positions(obj, offset)
}
\arguments{
\item{obj}{UnityObject with player log and positions data to be translated}

\item{offset}{New X, Z and Y vector. Needs to be a int vector of lenght 3 (X, Y, Z).}
}
\description{
Operates on given UnityObject, as it needs to fix not only player log, but potential goal positions in the object as well
}
