% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-getters.R
\name{was_trial_force_finished}
\alias{was_trial_force_finished}
\title{Returns if the trial has been force finished}
\usage{
was_trial_force_finished(obj, iTrial, zero_based = FALSE)
}
\arguments{
\item{obj}{\code{\link{BrainvrObject}}}

\item{iTrial}{trial index starting with 1 or 0 if \code{zero_based=TRUE}}

\item{zero_based}{if TRUE, it keeps the indices as they are reported
by the framework, beginning with 0. Otherwise trials start with 1}
}
\value{
logical
}
\description{
Returns if the trial has been force finished
}
