% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvr-getters.R
\name{was_trial_force_finished}
\alias{was_trial_force_finished}
\title{Returns if the trial has been force finished}
\usage{
was_trial_force_finished(obj, iTrial)
}
\arguments{
\item{obj}{BrainvrObject}

\item{iTrial}{trial index (starting with 1)}
}
\description{
Returns if the trial has been force finished
}
